section No.8 done

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video"; 
+import User from "../models/User"; 
 
 export const recommendedVideos = async (req, res) => {
     const dbVideos = await Video.find({})
@@ -24,21 +25,24 @@ export const searchVideos = async (req, res) => {
 
 export const watchVideo = async (req, res) => {
     const videoId = req.params.id;
-    const dbVideo = await Video.findById(videoId);
+    const dbVideo = await Video.findById(videoId).populate("owner");
     if (dbVideo === null) {
         return res.status(404).render("404");    
     }
     return res.render("videos/watch-video", {
         pageTitle: `Watch ${dbVideo.title}`,
         tabTitle: dbVideo.title,
-        dbVideo
+        dbVideo,
     });
 }
 export const getEdit = async (req, res) => {
     const videoId = req.params.id;
     const dbVideo = await Video.findById(videoId);
-    if (dbVideo === null) {
+    if (!dbVideo) {
         return res.status(404).render("404");
+    }
+    if (String(dbVideo.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("videos/edit-video", {
         pageTitle: `Edit ${dbVideo.title}`,
@@ -49,9 +53,13 @@ export const getEdit = async (req, res) => {
 export const postEdit = async (req, res) => {
     const videoId = req.params.id;
     const isDbVideoExists = await Video.exists({ _id: videoId });
+    const video = await Video.findById(videoId);
     const { title, hashtags, description } = req.body;
-    if (isDbVideoExists === false) {
+    if (!isDbVideoExists) {
         return res.render("404");
+    }
+    if (String(video.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(videoId, {  
         title,
@@ -62,6 +70,13 @@ export const postEdit = async (req, res) => {
 }
 export const deleteVideo = async (req, res) => {
     const { id } = req.params; 
+    const dbVideo = await Video.findById(id);
+    if (!dbVideo) {
+        return res.status(404).render("404");
+    }
+    if (String(dbVideo.owner) !== String(req.session.user._id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
@@ -70,17 +85,24 @@ export const getUploadVideo = (req, res) => {
 }
 export const postUploadVideo = async (req, res) => {
     const {
+        session: {
+            user: { _id }
+        }, 
         body: { title, description, hashtags },
         file: { path: fileUrl }
     } = req;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             fileUrl: fileUrl,
             title,
             description,
             hashtags: Video.formatHashtags(hashtags),
             createdAt: Date.now(),
+            owner: _id
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     } catch(error) {
         res.status(400).render("videos/upload-video", {
